requirements.txt
anyio==4.1.0
arrow==1.3.0
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
flet==0.15.0
flet-core==0.15.0
flet-runtime==0.15.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
Pygments==2.17.2
pyperclip==1.8.2
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
urllib3==2.1.0
watchdog==3.0.0
websocket-client==1.7.0
websockets==11.0.3
